refactor(middleware-cors) split checks and configure functions
