+++ packages/middleware-cors/src/checkRegexpOrigin.ts
@@ -1,0 +1,4 @@
+export const checkRegexpOrigin = (
+  origin: string,
+  option: string | string[] | RegExp,
+): boolean => (option instanceof RegExp && option.test(origin) ? true : false);
