+++ packages/middleware-cors/src/spec/index.spec.ts
@@ -1,4 +1,4 @@
-import { cors$ } from './index';
+import { cors$ } from '../index';
 
 describe('@marblejs/middleware-cors public API', () => {
   it('should be defined', () => {
